+++ app/com/lynxanalytics/biggraph/frontend_operations/Operations.scala
@@ -415,6 +415,100 @@ abstract class ProjectOperations(env: SparkFreeEnvironment) extends OperationReg
     val op = graph_operations.CreateStringScalar(data)
     op.result.created
   }
+
+  def checkTypeCollision(project: ProjectEditor, other: ProjectViewer) = {
+    val commonAttributeNames =
+      project.vertexAttributes.keySet & other.vertexAttributes.keySet
+    for (name <- commonAttributeNames) {
+      val a1 = project.vertexAttributes(name)
+      val a2 = other.vertexAttributes(name)
+      assert(
+        a1.typeTag.tpe =:= a2.typeTag.tpe,
+        s"Attribute '$name' has conflicting types in the two projects: " +
+          s"(${a1.typeTag.tpe} and ${a2.typeTag.tpe})")
+    }
+  }
+
+  def mergeInto(project: ProjectEditor, other: ProjectViewer): Unit = {
+    if (project.vertexSet == null) {
+      project.vertexSet = other.vertexSet
+      project.vertexAttributes = other.vertexAttributes
+      project.edgeBundle = other.edgeBundle
+      project.edgeAttributes = other.edgeAttributes
+      return
+    }
+
+    checkTypeCollision(project, other)
+    val vsUnion = {
+      val op = graph_operations.VertexSetUnion(2)
+      op(op.vss, Seq(project.vertexSet, other.vertexSet)).result
+    }
+
+    val newVertexAttributes = unifyAttributes(
+      project.vertexAttributes
+        .map {
+          case (name, attr) =>
+            name -> attr.pullVia(vsUnion.injections(0).reverse)
+        },
+      other.vertexAttributes
+        .map {
+          case (name, attr) =>
+            name -> attr.pullVia(vsUnion.injections(1).reverse)
+        })
+    val ebInduced = Option(project.edgeBundle).map { eb =>
+      val op = graph_operations.InducedEdgeBundle()
+      val mapping = vsUnion.injections(0)
+      op(op.srcMapping, mapping)(op.dstMapping, mapping)(op.edges, project.edgeBundle).result
+    }
+    val otherEbInduced = Option(other.edgeBundle).map { eb =>
+      val op = graph_operations.InducedEdgeBundle()
+      val mapping = vsUnion.injections(1)
+      op(op.srcMapping, mapping)(op.dstMapping, mapping)(op.edges, other.edgeBundle).result
+    }
+
+    val (newEdgeBundle, myEbInjection, otherEbInjection): (EdgeBundle, EdgeBundle, EdgeBundle) =
+      if (ebInduced.isDefined && !otherEbInduced.isDefined) {
+        (ebInduced.get.induced, ebInduced.get.embedding, null)
+      } else if (!ebInduced.isDefined && otherEbInduced.isDefined) {
+        (otherEbInduced.get.induced, null, otherEbInduced.get.embedding)
+      } else if (ebInduced.isDefined && otherEbInduced.isDefined) {
+        val idUnion = {
+          val op = graph_operations.VertexSetUnion(2)
+          op(
+            op.vss,
+            Seq(ebInduced.get.induced.idSet, otherEbInduced.get.induced.idSet))
+            .result
+        }
+        val ebUnion = {
+          val op = graph_operations.EdgeBundleUnion(2)
+          op(
+            op.ebs, Seq(ebInduced.get.induced.entity, otherEbInduced.get.induced.entity))(
+              op.injections, idUnion.injections.map(_.entity)).result.union
+        }
+        (
+          ebUnion,
+          idUnion.injections(0).reverse.concat(ebInduced.get.embedding),
+          idUnion.injections(1).reverse.concat(otherEbInduced.get.embedding))
+      } else {
+        (null, null, null)
+      }
+    val newEdgeAttributes = unifyAttributes(
+      project.edgeAttributes
+        .map {
+          case (name, attr) => name -> attr.pullVia(myEbInjection)
+        },
+      other.edgeAttributes
+        .map {
+          case (name, attr) => name -> attr.pullVia(otherEbInjection)
+        })
+
+    project.vertexSet = vsUnion.union
+    for ((name, attr) <- newVertexAttributes) {
+      project.newVertexAttribute(name, attr) // Clear notes.
+    }
+    project.edgeBundle = newEdgeBundle
+    project.edgeAttributes = newEdgeAttributes
+  }
 }
 
 object ScalaUtilities {
